+++ HISTORY.rst
@@ -5,6 +5,8 @@ Changelog
 ******************
 
 - Added ability to preview and undo commands during a recorder session.
+- ``{now}`` can be included in prompt.
+- Added 'pws' theme.
 
 
 2.0 (2014-06-21)
@@ -27,7 +29,7 @@ Changelog
 - Prompt variables can have ANSI colors and styles.
 - ``{hostname}`` can be included in prompt.
 - ``{git_branch}`` can be included in prompt.
-- Prompt variable ``{{full_cwd}}`` renamed to ``{{cwd}}``.
+- Prompt variable ``{full_cwd}`` renamed to ``{cwd}``.
 - Prompt variable ``{cwd}`` renamed to ``{dir}``.
 - Short option for ``--speed`` is now ``-s``.
 - Short option for ``--shell`` is now ``-S``.